packages/lidar-visualizer-backend/pyproject.toml
[project]
name = "lidar-api"
version = "0.1.0"
description = "LiDAR visualization API backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "laspy[laszip]>=2.5.0",
    "numpy>=1.21.0,<2.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "open3d>=0.17.0",
    "lazrs>=0.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lidar_api"]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]

[tool.uv]
config-settings-package = { numpy = { editable_mode = "compat" } }